fluid clean: remove fluid.ir to framework.ir and some funcs form fluid.layer.io to incubate.
